# internal/app/http/docs/swagger.yaml
basePath: /
definitions:
  dto.ErrorResponseDTO:
    properties:
      message:
        type: string
    type: object
  dto.UpsertUserSettingRequestDTO:
    properties:
      meuPluggyAccountIDs:
        items:
          type: string
        type: array
      meuPluggyClientID:
        type: string
      meuPluggyClientSecret:
        type: string
      notionPageID:
        type: string
      notionToken:
        type: string
    required:
    - meuPluggyAccountIDs
    - meuPluggyClientID
    - meuPluggyClientSecret
    - notionPageID
    - notionToken
    type: object
info:
  contact:
    email: danielmesquitta123@gmail.com
    name: Daniel Mesquita
  description: This API is responsible for syncing OpenFinance data to Notion.
  title: OpenFinance to Notion API
  version: "1.0"
paths:
  /auth/login/google:
    get:
      consumes:
      - application/json
      description: This endpoint is responsible for starting OAuth authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponseDTO'
      summary: BeginOAuth.
      tags:
      - Auth
  /to-notion:
    post:
      consumes:
      - application/json
      description: This endpoint is responsible for syncing OpenFinance data to Notion.
      parameters:
      - description: 'Start date (format RFC3339: 2006-01-02T15:04:05Z07:00)'
        in: query
        name: start_date
        type: string
      - description: 'End date (format RFC3339: 2006-01-02T15:04:05Z07:00)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponseDTO'
      summary: OpenFinance to Notion.
      tags:
      - Notion
  /users/me/settings:
    post:
      consumes:
      - application/json
      description: This endpoint is responsible for updating and creating user settings.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertUserSettingRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Upsert user setting.
      tags:
      - Setting
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"